backend/requirements.txt
langgraph>=0.2
langchain>=0.2
langchain-community>=0.2
langchain-openai>=0.1
openai>=1.40
psycopg[binary]>=3.2
SQLAlchemy>=2.0
langchain-postgres
pgvector>=0.2
python-dotenv>=1.0
pydantic>=2.6
langgraph-cli[inmem]>=0.1
pydantic>=2.6
fastapi
httpx
